Refactor: 시간표 검증 부분을 Util로 분리
UserTimetableService의 코드가 너무 길어지고 책임이 무거운 듯 해서 검증 부분만 Util로 분리
